tests reorganised
