+++ tests/test_integration/test_core/test_cli_end_to_end.py
