+++ tests/test_integration/test_core/test_template_manager_integration.py
