+++ tests/test_unit/test_core/test_cache.py
