+++ tests/test_unit/test_core/test_template_manager_unit.py
